Add mastra support
