--- typescript/examples/by-framework/mastra/index.ts
@@ -1,81 +1,0 @@
-import readline from "node:readline";
-
-import { openai } from "@ai-sdk/openai";
-import { Agent } from "@mastra/core";
-
-import { http } from "viem";
-import { createWalletClient } from "viem";
-import { privateKeyToAccount } from "viem/accounts";
-import { baseSepolia } from "viem/chains";
-
-import { getOnChainTools } from "@goat-sdk/adapter-mastra";
-import { PEPE, USDC, erc20 } from "@goat-sdk/plugin-erc20";
-
-import { sendETH } from "@goat-sdk/wallet-evm";
-import { viem } from "@goat-sdk/wallet-viem";
-
-require("dotenv").config();
-
-// 1. Create a wallet client
-const account = privateKeyToAccount(process.env.WALLET_PRIVATE_KEY as `0x${string}`);
-
-const walletClient = createWalletClient({
-    account: account,
-    transport: http(process.env.RPC_PROVIDER_URL),
-    chain: baseSepolia,
-});
-
-(async () => {
-    // 2. Get your onchain tools for your wallet
-    const tools = await getOnChainTools({
-        wallet: viem(walletClient),
-        plugins: [
-            sendETH(), // Enable ETH transfers
-            erc20({ tokens: [USDC, PEPE] }), // Enable ERC20 token operations
-        ],
-    });
-
-    // 3. Create an agent
-    const moneyTransmitterAgent = new Agent({
-        name: "Money Transmitter Agent",
-        instructions: `You are a money transmitter agent. You are responsible for sending money to the user's wallet.`,
-        model: openai("gpt-4o-mini"),
-        tools: tools,
-    });
-
-    // 4. Create a readline interface to interact with the agent
-    const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-    });
-
-    while (true) {
-        const prompt = await new Promise<string>((resolve) => {
-            rl.question('Enter your prompt (or "exit" to quit): ', resolve);
-        });
-
-        if (prompt === "exit") {
-            rl.close();
-            break;
-        }
-
-        console.log("\n-------------------\n");
-        console.log("TOOLS CALLED");
-        console.log("\n-------------------\n");
-        try {
-            const result = await moneyTransmitterAgent.generate(prompt, {
-                onStepFinish: (step) => {
-                    console.log(step);
-                },
-            });
-
-            console.log("\n-------------------\n");
-            console.log("RESPONSE");
-            console.log("\n-------------------\n");
-            console.log(result.text);
-        } catch (error) {
-            console.error(error);
-        }
-        console.log("\n-------------------\n");
-    }
-})();
