+++ typescript/examples/by-framework/mastra/package.json
@@ -1,6 +1,7 @@
 {
     "name": "goat-examples-mastra",
     "version": "0.0.0",
+    "type": "module",
     "description": "",
     "private": true,
     "scripts": {
@@ -17,8 +18,13 @@
         "@goat-sdk/wallet-viem": "0.2.12",
         "@mastra/core": "0.6.3",
         "dotenv": "^16.4.5",
-        "mastra": "0.4.2",
         "viem": "2.23.4",
         "zod": "3.23.8"
+    },
+    "devDependencies": {
+        "@types/node": "22.13.13",
+        "mastra": "0.4.2",
+        "tsx": "4.19.2",
+        "typescript": "5.8.2"
     }
 }
