+++ typescript/examples/by-framework/mastra/src/mastra/index.ts
@@ -1,0 +1,7 @@
+import { Mastra } from "@mastra/core";
+
+import { moneyTransmitter } from "./agents/moneyTransmitter";
+
+export const mastra = new Mastra({
+    agents: { moneyTransmitter },
+});
