+++ typescript/examples/by-framework/mastra/tsconfig.json
@@ -1,8 +1,14 @@
 {
-    "extends": "../../../tsconfig.base.json",
     "compilerOptions": {
+        "target": "ES2022",
+        "module": "ES2022",
+        "moduleResolution": "bundler",
+        "esModuleInterop": true,
+        "forceConsistentCasingInFileNames": true,
+        "strict": true,
+        "skipLibCheck": true,
         "outDir": "dist"
     },
-    "include": ["index.ts"],
-    "exclude": ["node_modules", "dist"]
+    "include": ["src/**/*"],
+    "exclude": ["node_modules", "dist", ".mastra"]
 }
