+++ typescript/packages/adapters/mastra/src/index.ts
@@ -1,6 +1,7 @@
 import { type GetToolsParams, type ToolBase, type WalletClientBase, getTools } from "@goat-sdk/core";
-import type { Tool } from "@mastra/core";
-import { createTool } from "@mastra/core";
+
+import { type CoreTool, tool } from "ai";
+import type { z } from "zod";
 
 export type GetOnChainToolsParams<TWalletClient extends WalletClientBase> = GetToolsParams<TWalletClient>;
 
@@ -13,24 +14,17 @@ export async function getOnChainTools<TWalletClient extends WalletClientBase>({
         plugins,
     });
 
-    const mastraTools = new Map<string, Tool>();
+    const aiTools: { [key: string]: CoreTool } = {};
 
     for (const t of tools) {
-        mastraTools.set(
-            t.name,
-            createTool({
-                id: t.name,
-                description: t.description,
-                // biome-ignore lint/suspicious/noExplicitAny: need to fix zod versions
-                inputSchema: t.parameters as any,
-                execute: async ({ context }) => {
-                    // Validate input against original schema
-                    const validatedInput = await t.parameters.parseAsync(context);
-                    return await t.execute(validatedInput);
-                },
-            }),
-        );
+        aiTools[t.name] = tool({
+            description: t.description,
+            parameters: t.parameters,
+            execute: async (arg: z.output<typeof t.parameters>) => {
+                return await t.execute(arg);
+            },
+        });
     }
 
-    return mastraTools;
+    return aiTools;
 }
